requirements.txt
langchain
langchain-community
chromadb
pypdf
streamlit
google-generativeai
python-dotenv